requirements.txt
bigdl
apache-airflow
pandas
kafka-python
scikit-learn
torch